Refactor(web): Restructure header module organization
Moved the header and its related components into a new "shell"
directory to improve module organization. Updated import paths in all
 affected files to reflect the changes. This refactor aims to enhance
  the maintainability and clarity of the header module structure.
