+++ apps/web/app/shell/header/header.tsx
@@ -6,8 +6,8 @@ import { SuddenlyGiovanni } from '@suddenlygiovanni/ui/components/suddenly-giova
 import { useToggle } from '@suddenlygiovanni/ui/hooks/use-toggle.tsx'
 import { clsx } from '@suddenlygiovanni/ui/lib/utils.ts'
 
-import avatarAssetUrl from '../assets/giovanni_ravalico-profile_bw.webp'
-import { routesRecord } from '../routes-record.ts'
+import avatarAssetUrl from '../../assets/giovanni_ravalico-profile_bw.webp'
+import { routesRecord } from '../../routes-record.ts'
 import { NavLink } from './nav-link.tsx'
 import { ThemeSwitch } from './theme-switch.tsx'
